change state.load_options to a tuple
having this be an immutable sequence is safer
and possibly lower overhead.

The change here went in with no issues save
for tests that asserted it was a set.
InstanceState.load_options is only referred
towards when the object is first loaded, and then
within the logic that emits an object refresh
as well as within a lazy loader.   it's only
accessed as a whole collection.

Fixes: #7558
Change-Id: Id1adbec0f93bcfbfc934ec9cd39e71e74727845d