repair broken truediv test suite; memusage
the truediv test suite didn't have __backend__ so wasn't running
for every DB except in the main build.  Repaired this as well
as truediv support to preserve the right-hand side type
when casting to numeric, if the right type is already a
numeric type.

also fixed a memusage test that relies on savepoints so was
not running under gerrit runs.

Change-Id: I3be223fdf697af9c1ed61b70d621f57cbbb7a92b