correct for non-deterministic gc artifacts
Observed the tests here have different profiling
counts when run individually vs. as a group, and
this seems to be due to whether or not results of
each query are garbage collected or not.  for
all but one test, ensuring results stay between
query runs seems to meet the current profiling
counts.

Change-Id: I5aca5db08936757ad2a6055c5fc077cc58979bdd